Flags for features and releng imports

+++ src/index_runner/main.py
@@ -128,7 +128,8 @@ def _handle_msg(msg):
     if event_type in ['REINDEX', 'NEW_VERSION', 'COPY_OBJECT', 'RENAME_OBJECT']:
         obj = _fetch_obj_data(msg)
         ws_info = _fetch_ws_info(msg)
-        releng_importer.run_importer(obj, ws_info, msg)
+        if not config()['skip_releng']:
+            releng_importer.run_importer(obj, ws_info, msg)
         es_indexer.run_indexer(obj, ws_info, msg)
     elif event_type == 'REINDEX_WS' or event_type == 'CLONE_WORKSPACE':
         # Reindex all objects in a workspace, overwriting existing data
@@ -145,7 +146,7 @@ def _handle_msg(msg):
             obj_ref = f"{msg['wsid']}/{msg['objid']}/{msg.get('ver', '?')}"
             obj = _fetch_obj_data(msg)
             ws_info = _fetch_ws_info(msg)
-            if not exists_in_releng:
+            if not exists_in_releng and not config()['skip_releng']:
                 logger.info(f"Importing object {obj_ref} into RE.")
                 releng_importer.run_importer(obj, ws_info, msg)
             if not exists_in_es:
@@ -154,15 +155,18 @@ def _handle_msg(msg):
     elif event_type == 'OBJECT_DELETE_STATE_CHANGE':
         # Delete the object on RE and ES. Synchronous for now.
         es_indexer.delete_obj(msg)
-        releng_importer.delete_obj(msg)
+        if not config()['skip_releng']:
+            releng_importer.delete_obj(msg)
     elif event_type == 'WORKSPACE_DELETE_STATE_CHANGE':
         # Delete everything in RE and ES under this workspace
         es_indexer.delete_ws(msg)
-        releng_importer.delete_ws(msg)
+        if not config()['skip_releng']:
+            releng_importer.delete_ws(msg)
     elif event_type == 'SET_GLOBAL_PERMISSION':
         # Set the `is_public` permissions for a workspace
         es_indexer.set_perms(msg)
-        releng_importer.set_perms(msg)
+        if not config()['skip_releng']:
+            releng_importer.set_perms(msg)
     elif event_type == 'RELOAD_ELASTIC_ALIASES':
         # Reload aliases on ES from the global config file
         es_indexer.reload_aliases()
